Added functionality + minor design tweaks
I removed the second button, as the first button is able to handle all
conditions.  Added functionality for when selecting your age.  If you
are under 18, the conditionsLabel (set to invisible) will change to a
message for those under the age, and will become visible.
Alternatively, if the age is set to over 18, the rest of the form will
become visible so that they can continue to fill out the form.

Added functionality to the form and buttons.  The submitButton now uses
a sequens of if statements to determine which of the 3 forms to open.